+++ app/queries/servers/channel.ts
@@ -466,16 +466,21 @@ export function observeMyChannelMentionCount(database: Database, teamId?: string
 
 export function queryMyChannelsByUnread(database: Database, isUnread: boolean, sortBy: 'last_viewed_at' | 'last_post_at', take: number, excludeIds?: string[]) {
     const clause: Q.Clause[] = [Q.where('is_unread', Q.eq(isUnread))];
+    const count: Q.Clause[] = [];
 
     if (excludeIds?.length) {
         clause.push(Q.where('id', Q.notIn(excludeIds)));
+    }
+
+    if (take > 0) {
+        count.push(Q.take(take));
     }
 
     return queryAllMyChannel(database).extend(
         Q.on(CHANNEL, Q.where('delete_at', Q.eq(0))),
         ...clause,
         Q.sortBy(sortBy, Q.desc),
-        Q.take(take),
+        ...count,
     );
 }
