+++ app/screens/home/channel_list/categories_list/categories/unreads/index.ts
@@ -8,6 +8,7 @@ import {combineLatestWith, map, switchMap} from 'rxjs/operators';
 
 import {Preferences} from '@constants';
 import {getPreferenceAsBool} from '@helpers/api/preference';
+import {filterAndSortMyChannels, makeChannelsMap} from '@helpers/database';
 import {getChannelById, observeChannelsByLastPostAt, observeNotifyPropsByChannels, queryMyChannelUnreads} from '@queries/servers/channel';
 import {queryPreferencesByCategoryAndName} from '@queries/servers/preference';
 import {observeLastUnreadChannelId} from '@queries/servers/system';
@@ -16,7 +17,6 @@ import UnreadCategories from './unreads';
 
 import type {WithDatabaseArgs} from '@typings/database/database';
 import type ChannelModel from '@typings/database/models/servers/channel';
-import type MyChannelModel from '@typings/database/models/servers/my_channel';
 import type PreferenceModel from '@typings/database/models/servers/preference';
 
 type WithDatabaseProps = WithDatabaseArgs & {
@@ -30,66 +30,8 @@ type CA = [
     b: ChannelModel | undefined,
 ]
 
-type NotifyProps = {
-    [key: string]: Partial<ChannelNotifyProps>;
-}
-
-/**
- * Filtering / Sorting:
- *
- * Unreads, Mentions, and Muted Mentions Only
- * Mentions on top, then unreads, then muted channels with mentions.
- */
-
- type FilterAndSortMyChannelsArgs = [
-    MyChannelModel[],
-    Record<string, ChannelModel>,
-    NotifyProps,
-]
-
 const concatenateChannelsArray = ([a, b]: CA) => {
     return of$(b ? a.filter((c) => c && c.id !== b.id).concat(b) : a);
-};
-
-const filterAndSortMyChannels = ([myChannels, channels, notifyProps]: FilterAndSortMyChannelsArgs): ChannelModel[] => {
-    const mentions: ChannelModel[] = [];
-    const unreads: ChannelModel[] = [];
-    const mutedMentions: ChannelModel[] = [];
-
-    const isMuted = (id: string) => {
-        return notifyProps[id]?.mark_unread === 'mention';
-    };
-
-    for (const myChannel of myChannels) {
-        const id = myChannel.id;
-
-        // is it a mention?
-        if (!isMuted(id) && myChannel.mentionsCount > 0 && channels[id]) {
-            mentions.push(channels[id]);
-            continue;
-        }
-
-        // is it unread?
-        if (!isMuted(myChannel.id) && myChannel.isUnread && channels[id]) {
-            unreads.push(channels[id]);
-            continue;
-        }
-
-        // is it a muted mention?
-        if (isMuted(myChannel.id) && myChannel.mentionsCount > 0 && channels[id]) {
-            mutedMentions.push(channels[id]);
-            continue;
-        }
-    }
-
-    return [...mentions, ...unreads, ...mutedMentions];
-};
-
-const makeChannelsMap = (channels: ChannelModel[]) => {
-    return channels.reduce<Record<string, ChannelModel>>((result, c) => {
-        result[c.id] = c;
-        return result;
-    }, {});
 };
 
 const enhanced = withObservables(['currentTeamId', 'isTablet', 'onlyUnreads'], ({currentTeamId, isTablet, database, onlyUnreads}: WithDatabaseProps) => {
